requirements.txt
streamlit
sentence-transformers
nltk
numpy
scikit-learn
requests
beautifulsoup4
faiss-cpu
PyMuPDF
